+++ src/components/modal/components/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Modal components module
+ * Exports all modal-related components for the Flowboard application
+ */
+
+export { ModalHeader } from './ModalHeader';
+export { ModalFormSection } from './ModalFormSection';
+export { ModalLabelSection } from './ModalLabelSection';
+export { ModalChecklistSection } from './ModalChecklistSection';
+export { EnhancedChecklistManager } from './EnhancedChecklistManager';